+++ src/Tests/Unit/Catesta-Module.Tests.ps1
@@ -7,8 +7,9 @@ $PathToModule = [System.IO.Path]::Combine('..', '..', $ModuleName, "$ModuleName.
 #-------------------------------------------------------------------------
 Describe 'Module Tests' -Tag Unit {
     Context "Module Tests" {
+        $script:manifestEval = $null
         It 'Passes Test-ModuleManifest' {
-            Test-ModuleManifest -Path $PathToManifest | Should Not BeNullOrEmpty
+            { $script:manifestEval = Test-ModuleManifest -Path $PathToManifest } | Should Not throw
             $? | Should Be $true
         }#manifestTest
         It 'root module Catesta.psm1 should exist' {
@@ -16,8 +17,22 @@ Describe 'Module Tests' -Tag Unit {
             $? | Should Be $true
         }#psm1Exists
         It 'manifest should contain Catesta.psm1' {
-            $PathToManifest |
-                Should -FileContentMatchExactly "Catesta.psm1"
+            $PathToManifest | Should -FileContentMatchExactly "Catesta.psm1"
         }#validPSM1
+        It 'should have a matching module name in the manifest' {
+            $script:manifestEval.Name | Should Be $ModuleName
+        }#name
+        It 'should have a valid description in the manifest' {
+            $script:manifestEval.Description | Should Not BeNullOrEmpty
+        }#description
+        It 'should have a valid author in the manifest' {
+            $script:manifestEval.Author | Should Not BeNullOrEmpty
+        }#author
+        It 'should have a valid version in the manifest' {
+            $script:manifestEval.Version -as [Version] | Should Not BeNullOrEmpty
+        }#version
+        It 'should have a valid guid in the manifest' {
+            { [guid]::Parse($script:manifestEval.Guid) } | Should Not throw
+        }#guid
     }#context_ModuleTests
 }#describe_ModuleTests